Fix race where sending messages after large messages may fail
With large messages on XHR-streaming and XHR-poll (possibly others)
the stateHTTPReceiverClosed state may be set on httpReceiver, which
causes future calls to sendBulk to be a no-op until another receiver
attaches. This is usually not a problem because a goroutine in
session.attachReceiver calls detachReceiver soon after, and detach
causes the session's recv field to be nulled out. However, if msgs
are sent between the time when stateHTTPReceiverClosed is called
and detachReceiver occurs, then those messages will be silently
ignored.